// src/app/modules/ConsultaEnvio/model/plataforma.ts
export class PlataForma {
    ID: number;
    TRANSPORTISTA: string;
    CODIGO_OFICINA: string;
    NOMBRE_OFICINA: string;
    DIRECCION_OFICINA: string;
    POBLACION_OFICINA: string;
    CP_OFICINA: number;
    PROVINCIA: string;
    TELEFONO: string;
    EMAIL: string;
    HORARIO: string;
    FAX: string;
    OBSERVACIONES: string;
    PLAZA: string;
    NUMERO_ENVIO_CORREOS: string;

    constructor () {}
}